Split RCN tests and without_variable_measure() tests
